refactor(plugins): flatten googledrive route module indirection
Address reviewer feedback on misleading indirection in the route layer:

- Move _sync_folder_files from route_utils.py into folders.py, its only
  consumer. The leading underscore was misleading while the function was
  imported cross-module; it is now genuinely module-private. route_utils.py
  is left holding only get_drive_credentials, which is shared by all three
  route modules.
- Drop the get_valid_access_token re-export from dependencies.py. It is a
  directly-called helper, not a FastAPI dependency, so oauth.py, folders.py
  and files.py now import it straight from app.core_plugins.googledrive.oauth.
