+++ app/core_plugins/googledrive/routes/files.py
@@ -15,7 +15,8 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from app.core.db import get_async_session, get_session
 from app.core_plugins.googledrive.client import GoogleDriveClient
 from app.core_plugins.googledrive.constants import DRIVE_MAX_UPLOAD_BYTES
-from app.core_plugins.googledrive.routes.dependencies import get_valid_access_token, require_user_id
+from app.core_plugins.googledrive.oauth import get_valid_access_token
+from app.core_plugins.googledrive.routes.dependencies import require_user_id
 from app.core_plugins.googledrive.routes.route_utils import get_drive_credentials
 from app.core_plugins.googledrive.types import (
     DriveBrowseItem,
