+++ app/core_plugins/googledrive/routes/oauth.py
@@ -15,10 +15,11 @@ from app.core_plugins.googledrive.oauth import (
     generate_authorization_url,
     get_token_record,
     get_user_info,
+    get_valid_access_token,
     revoke_token,
     save_tokens,
 )
-from app.core_plugins.googledrive.routes.dependencies import get_valid_access_token, require_user_id
+from app.core_plugins.googledrive.routes.dependencies import require_user_id
 from app.core_plugins.googledrive.routes.route_utils import get_drive_credentials
 from app.core_plugins.googledrive.types import AuthorizationUrlResponse, ConnectionStatusResponse
 from app.lib.log import get_logger
