+++ app/core_plugins/googledrive/tests/test_routes.py
@@ -227,7 +227,6 @@ class TestSyncFolder:
         with (
             patch("app.core_plugins.googledrive.routes.folders.GoogleDriveClient") as mock_client_cls,
             patch("app.core_plugins.googledrive.routes.folders.process_folder_rag", new_callable=AsyncMock),
-            patch("app.core_plugins.googledrive.routes.route_utils.GoogleDriveClient") as mock_route_utils_client_cls,
         ):
             mock_client = AsyncMock()
             mock_client.get_folder.return_value = folder_metadata
@@ -235,8 +234,6 @@ class TestSyncFolder:
             mock_client.__aenter__.return_value = mock_client
             mock_client.__aexit__.return_value = None
             mock_client_cls.return_value = mock_client
-            mock_route_utils_client_cls.return_value = mock_client
-            mock_route_utils_client_cls.FOLDER_MIME_TYPE = "application/vnd.google-apps.folder"
 
             response = await drive_client.post(
                 "/api/v1/googledrive/folders/sync",
@@ -276,7 +273,7 @@ class TestRefreshFolder:
     ) -> None:
         """POST /folders/{id}/refresh should re-sync files from Drive."""
         with (
-            patch("app.core_plugins.googledrive.routes.route_utils.GoogleDriveClient") as mock_client_cls,
+            patch("app.core_plugins.googledrive.routes.folders.GoogleDriveClient") as mock_client_cls,
             patch("app.core_plugins.googledrive.routes.folders.process_folder_rag", new_callable=AsyncMock),
         ):
             mock_client = AsyncMock()
@@ -312,7 +309,7 @@ class TestRefreshFolder:
     ) -> None:
         """Subfolder entries returned by the Drive API must not be stored as DriveFiles."""
         with (
-            patch("app.core_plugins.googledrive.routes.route_utils.GoogleDriveClient") as mock_client_cls,
+            patch("app.core_plugins.googledrive.routes.folders.GoogleDriveClient") as mock_client_cls,
             patch("app.core_plugins.googledrive.routes.folders.process_folder_rag", new_callable=AsyncMock),
         ):
             mock_client = AsyncMock()
